Restrict AddCommand configuration to the WorkingTreeIterator.
The AddCommand has only three configurations: the files to add (which should be handled via FileSelector#getSelectedFiles(); the 'Update' option, which prevents new files from being tracked and is thus an undesired option; and the 'WorkingTreeIterator' option, which is the only option that might be customized. Though it is probable that the WorkingTreeIterator option will only be useful if uses a different approach to running the JGit code, it's still provided in case that isn't the case.